requirements.txt
streamlit==1.37.1
pandas==2.2.2
python-dotenv==1.0.1
plotly==5.23.0
